Peg version of io.js to 2.5.0 for travis
Currently issue #173 (https://github.com/brianmcd/contextify/issues/173) of contextify is causing all PRs to fail on io.js.
Until this issue is resolved, I'd recommend pinning the io.js version to v2.5.0 (the last in the 2.x line), such that we get accurate status from travis.

This can be reverted when the issue with contextify is resolved.