Add conditional scrolling into view of focused option
Previously the focused option would be scrolled to
the top of the scroll window when the menu is opened,
regardless of whether it was already in view.  An
example of a situation where this is problematic is
when a list starts with one or more disabled items.
In this situation, the disabled items are scrolled
out of view when the menu opens, and because they're
at the top of the list it's hard to notice they're
there.

This commit modifies the behavior such that when the
menu is opened, it will only scroll the focused option
into view if it's not already fully in view.